Changed ConfigLoader, and CachedConfigField
CachedConfigField<T> now has a new Constructor that takes in a class to fix the problem with getting `CraftWorld` with `World` and such. And for ConfigLoader, it has some more methods for the public to have access too. I will be editing this class a lot for lots of customizability with loading configs and such.